Updated Tile Added Class Texture Update Collision Processing Added Pickuping items

+++ OpenGL/src/Coords.h
@@ -12,8 +12,8 @@ Prosty obiekt z koordynatami
 struct Coords
 {
 public:
-	float posX;
-	float posY;
+	double posX;
+	double posY;
 	float Velocity;
 };
